File I/O!!!
Fixed a mistake that caused player gold to be stored as an int, not  a double (which in turn caused orc players to find 0-gold hoards).

+++ object.cc
@@ -38,7 +38,7 @@ posn posn::dirAdjacent(int dir)
 		p.row = row + 1, p.col = col - 1;
 	else if (dir == SOUTHEAST)
 		p.row = row + 1, p.col = col + 1;
-	else
+	else // south
 		p.row = row + 1, p.col = col;
 	return p;
 }
@@ -288,7 +288,7 @@ void Player::initRace(char c)
 }
 void Player::addGold(double value)
 {
-	int collected;
+	double collected;
 	if (pRace == DWARF)
 		collected = 2 * value;
 	else if (pRace == ORC)
